// src/app/utils/enum/usuarios.enum.ts
//ENUMS para poder ser usadas nas actions
export const enum usuariosTypeActions {
  LOAD_USUARIOS = '[LOAD_USUARIOS]: CARREGANDO USUARIOS',
  LOAD_USUARIOS_SUCCESS = '[LOAD_USUARIOS_SUCCESS]: CARREGOU USUARIOS',
  LOAD_USUARIOS_FAIL = '[LOAD_USUARIOS_FAIL]: FALHOU USUARIOS',

  LOAD_USUARIO = '[LOAD_USUARIO]: CARREGANDO USUARIO',
  LOAD_USUARIO_SUCCESS = '[LOAD_USUARIO_SUCCESS]: CARREGOU USUARIO',
  LOAD_USUARIO_FAIL = '[LOAD_USUARIO_FAIL]: FALHOU USUARIO',

  CREATE_USUARIO = '[CREATE_USUARIO]: CRIANDO USUARIO',
  CREATE_USUARIO_SUCCESS = '[CREATE_USUARIO_SUCCESS]: CRIOU USUARIO',
  CREATE_USUARIO_FAIL = '[CREATE_USUARIO_FAIL]: FALHOU CRIAR USUARIO',

  UPDATE_USUARIO = '[UPDATE_USUARIO]: EDITANDO USUARIO',
  UPDATE_USUARIO_SUCCESS = '[UPDATE_USUARIO_SUCCESS]: EDITOU USUARIO',
  UPDATE_USUARIO_FAIL = '[UPDATE_USUARIO_FAIL]: FALHOU EDITAR USUARIO',

  DELETE_USUARIO = '[DELETE_USUARIO]: DELETANDO USUARIO',
  DELETE_USUARIO_SUCCESS = '[DELETE_USUARIO_SUCCESS]: DELETOU USUARIO',
  DELETE_USUARIO_FAIL = '[DELETE_USUARIO_FAIL]: FALHOU DELETAR USUARIO',
}